Splits tests into three files, adds error tests
